Fix parity audit compile errors (JToken blend, UVChannelFlags)
Use AddCompareRow for blending field instead of CmpScalar with JToken.
Cast TSA uvChannelMask to int for JSON serialization.

+++ tools/unity-quarks-exporter/Editor/UnityEffectProbe.cs
@@ -276,7 +276,7 @@ namespace BabylonQuarks.UnityExporter
                 .Set("cycleCount", tsa.cycleCount)
                 .Set("rowMode", tsa.rowMode.ToString())
                 .Set("rowIndex", tsa.rowIndex)
-                .Set("uvChannelMask", tsa.uvChannelMask)
+                .Set("uvChannelMask", (int)tsa.uvChannelMask)
                 .Set("timeMode", tsa.timeMode.ToString())
                 .Set("fps", tsa.fps)
                 .Set("startFrame", ProbeCurve(tsa.startFrame))
